datatables integration

+++ app/Http/Controllers/PostController/Update.php
@@ -14,42 +14,50 @@ use Illuminate\Support\Facades\File;
 class Update extends Controller
 {
     public function index(Post $post){
-        $countries = Country::get();
-        return view('user.post.update',['countries'=>$countries, 'post'=>$post]);
+        if($post->user_id == auth()->user()->id){
+            $countries = Country::get();
+            return view('user.post.update',['countries'=>$countries, 'post'=>$post]);
+        }
+        abort(403);
+        
     }
 
     public function update(Post $post, UpdatePostRequest $request){
-        $post->update($request->validated());
+        if($post->user_id == auth()->user()->id){
+            $post->update($request->validated());
 
-        $features = ['garage', 'security', 'parking', 'balcony', 'swimming_pool', 'garden', 'terrace'];
-        foreach($features as $feature){
-            if(!$request->has($feature)){
-                $post->$feature = false;
-                $post->save();
+            $features = ['garage', 'security', 'parking', 'balcony', 'swimming_pool', 'garden', 'terrace'];
+            foreach($features as $feature){
+                if(!$request->has($feature)){
+                    $post->$feature = false;
+                    $post->save();
+                }
             }
-        }
 
-        if(isset($request->media)){
-            foreach($request->media as $image){
-            
-                $from = public_path('tmp/uploads/'.$image);
-                $to = public_path('post_images/'.$image);
+            if(isset($request->media)){
+                foreach($request->media as $image){
                 
-                File::move($from, $to);
-                $post->images()->create([
-                    'name' => $image,
-                ]);
+                    $from = public_path('tmp/uploads/'.$image);
+                    $to = public_path('post_images/'.$image);
+                    
+                    File::move($from, $to);
+                    $post->images()->create([
+                        'name' => $image,
+                    ]);
+                }
             }
-        }
-        
-        if(isset($request->deleted_media)){
-            foreach($request->deleted_media as $deleted_media){
-                File::delete(public_path('post_images/'.$deleted_media));
-                Image::where('name', $deleted_media)->delete();
+            
+            if(isset($request->deleted_media)){
+                foreach($request->deleted_media as $deleted_media){
+                    File::delete(public_path('post_images/'.$deleted_media));
+                    Image::where('name', $deleted_media)->delete();
+                }
             }
+            
+            $countries = Country::get();
+            return redirect()->route('post.update',['countries'=>$countries, 'post'=>$post]);
         }
+        abort(403);
         
-        $countries = Country::get();
-        return redirect()->route('post.update',['countries'=>$countries, 'post'=>$post]);
     }
 }
